requirements.txt
fastapi==0.122.0
uvicorn==0.38.0
httpx==0.28.1
beautifulsoup4==4.14.2
python-dateutil==2.9.0.post0
